backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--no-binary psycopg2

apipkg==1.5               # via execnet
asgiref==3.2.3            # via django
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
beautifulsoup4==4.8.1     # via django-livereload-server
django-livereload-server==0.3.2
git+git://github.com/TracyWebTech/django-revproxy@master#egg=django-revproxy>=0.9.16
django==3.0b1
djangorestframework==3.10.3
execnet==1.7.1            # via pytest-xdist
gunicorn==20.0.4
importlib-metadata==0.23  # via pluggy, pytest
more-itertools==7.2.0     # via pytest, zipp
packaging==19.2           # via pytest
pluggy==0.13.0            # via pytest
psycopg2==2.8.4
py==1.8.0                 # via pytest
pyparsing==2.4.2          # via packaging
pytest-django==3.6.0
pytest-forked==1.1.3      # via pytest-xdist
pytest-xdist==1.30.0
pytest==5.2.1
pytz==2019.3              # via django
six==1.12.0               # via django-livereload-server, packaging, pytest-xdist
soupsieve==1.9.4          # via beautifulsoup4
sqlparse==0.3.0           # via django
tornado==6.0.3            # via django-livereload-server
urllib3==1.25.6
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata
